requirements.txt
flask
sqlmodel-controller[all]
pytest
pyjwt
python-dotenv
pytest-mock